# server/common/api.yml
openapi: 3.0.1
info:
  title: express-mongoose-template
  description: express-mongoose-template
  version: 1.0.0
servers:
  - url: /api/v1
...
paths:
...
  /examples/{id}:
    get:
      tags:
        - Examples
      parameters:
        - name: id
          in: path
          description: The id of the example to retrieve
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Return the example with the specified id
          content: {}
        404:
          description: Example not found
          content: {}
...
  /boms/{id}:
    delete:
      tags:
        - BOMs
      parameters:
        - name: id
          in: path
          description: The id of the BOM to delete
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Return the number of BOMs deleted
          content: {}
        404:
          description: BOM not found or not deleted
          content: {}
    get:
      tags:
        - BOMs
      parameters:
        - name: id
          in: path
          description: The id of the BOM to retrieve
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Return the BOM with the specified id
          content: {}
        404:
          description: BOM not found
          content: {}
    put:
      tags:
        - BOMs
      parameters:
        - name: id
          in: path
          description: The id of the BOM to update
          required: true
          schema:
            type: integer
      requestBody:
        description: a BOM
        content:
          application/json:
            schema:
              type: object
        required: true
      responses:
        200:
          description: Return the BOM with the specified id
          content: {}
        404:
          description: BOM not found
          content: {}
...
  /inventoryStmts/{id}:
    get:
      tags:
        - InventoryStmts
      parameters:
        - name: id
          in: path
          description: The id of the example to retrieve
          required: true
          schema:
            type: string
      responses:
        200:
          description: Return the inventoryStmt with the specified id
          content: {}
        404:
          description: inventoryStmt not found
          content: {}
    put:
      tags:
        - InventoryStmts
      parameters:
        - name: id
          in: path
          description: The id of the inventoryStmt to update
          required: true
          schema:
            type: integer
      requestBody:
        description: an inventoryStmt
        content:
          application/json:
            schema:
              type: object
        required: true
      responses:
        200:
          description: Return the inventoryStmt with the specified id
          content: {}
        404:
          description: inventoryStmt not found
          content: {}
...